% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Identifiable2.R
\name{Check_Identifiable2}
\alias{Check_Identifiable2}
\title{Check for identifiability of fixed effects}
\usage{
Check_Identifiable2(obj)
}
\arguments{
\item{obj, }{The compiled object}
}
\value{
A tagged list of the hessian and the message
}
\description{
\code{Check_Identifiable} calculates the matrix of second-derivatives of the marginal likelihood w.r.t. fixed effects, to see if any linear combinations are unidentifiable
}
