% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplot.R
\name{eplot}
\alias{eplot}
\title{Empty figure}
\usage{
eplot(...)
}
\arguments{
\item{base_size}{default text size for figure, default=14}

\item{base_family}{font type for figure, default=""}
}
\value{
empty plot
}
\description{
\code{eplot} Plot empty figure to fill with base plotting (e.g. lines, polygons, points, etc.)
}
\author{
M.B. Rudd
}
