\name{RuddR-package}
\alias{RuddR-package}
\alias{RuddR}
\docType{package}
\title{
\packageTitle{RuddR}
}
\description{
\packageDescription{RuddR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RuddR}
\packageIndices{RuddR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{RuddR}

Maintainer: \packageMaintainer{RuddR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
